# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AllergenAssessmentReason": {
        "enum": [
          "FEATURE_DISABLED",
          "REFERENCE_UNAVAILABLE",
          "EVIDENCE_UNAVAILABLE",
          "ASSESSMENT_FAILED"
        ],
        "title": "AllergenAssessmentReason",
        "type": "string"
      },
      "AllergenAssessmentResponse": {
        "properties": {
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/AllergenConceptOutcomeResponse"
            },
            "title": "Concepts",
            "type": "array"
          },
          "engine_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Version"
          },
          "evidence_coverage": {
            "$ref": "#/components/schemas/EvidenceCoverageState"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/AllergenFindingResponse"
            },
            "title": "Findings",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AllergenAssessmentReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "reference_dataset_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssessmentReferenceDatasetVersionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_signals": {
            "items": {
              "$ref": "#/components/schemas/PackageMatchEvidenceResponse"
            },
            "title": "Source Signals",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/AllergenAssessmentStatus"
          }
        },
        "required": [
          "status",
          "reason",
          "evidence_coverage"
        ],
        "title": "AllergenAssessmentResponse",
        "type": "object"
      },
      "AllergenAssessmentStatus": {
        "enum": [
          "COMPLETED",
          "NOT_ASSESSED"
        ],
        "title": "AllergenAssessmentStatus",
        "type": "string"
      },
      "AllergenConceptOutcomeResponse": {
        "properties": {
          "concept_id": {
            "title": "Concept Id",
            "type": "string"
          },
          "finding_ids": {
            "items": {
              "type": "string"
            },
            "title": "Finding Ids",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outcome": {
            "$ref": "#/components/schemas/AllergenAssessmentOutcome"
          },
          "parent_ids": {
            "items": {
              "type": "string"
            },
            "title": "Parent Ids",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "rule_ids": {
            "items": {
              "type": "string"
            },
            "title": "Rule Ids",
            "type": "array"
          }
        },
        "required": [
          "concept_id",
          "name",
          "outcome"
        ],
        "title": "AllergenConceptOutcomeResponse",
        "type": "object"
      },
      "AllergenFindingResponse": {
        "properties": {
          "concept_id": {
            "title": "Concept Id",
            "type": "string"
          },
          "end_index": {
            "title": "End Index",
            "type": "integer"
          },
          "engine_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Version"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "mapping_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapping Id"
          },
          "matched_text": {
            "title": "Matched Text",
            "type": "string"
          },
          "off_dataset_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Off Dataset Version Id"
          },
          "reference_dataset_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Dataset Version Id"
          },
          "relationship_type": {
            "$ref": "#/components/schemas/AllergenRelationshipType"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "source_field": {
            "title": "Source Field",
            "type": "string"
          },
          "source_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Revision"
          },
          "source_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Text"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "start_index": {
            "title": "Start Index",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "concept_id",
          "relationship_type",
          "matched_text",
          "start_index",
          "end_index",
          "source_field",
          "source_url"
        ],
        "title": "AllergenFindingResponse",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "IDENTIFIER_REQUIRED",
          "IDENTIFIER_CHARACTERS_INVALID",
          "IDENTIFIER_LENGTH_UNSUPPORTED",
          "IDENTIFIER_CHECK_DIGIT_INVALID",
          "PACKAGE_MATCH_SOURCE_UNAVAILABLE",
          "REFERENCE_IMAGE_URL_INVALID",
          "REFERENCE_IMAGE_NOT_FOUND",
          "REFERENCE_IMAGE_SOURCE_UNAVAILABLE",
          "RATE_LIMIT_EXCEEDED",
          "SEARCH_QUERY_REQUIRED",
          "SEARCH_QUERY_INVALID",
          "SEARCH_RESULTS_NOT_FOUND"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "EvidenceCoverageState": {
        "enum": [
          "NOT_ASSESSED",
          "COMPLETE_READABLE_LABEL",
          "PARTIAL",
          "UNREADABLE"
        ],
        "title": "EvidenceCoverageState",
        "type": "string"
      },
      "ExternalDatasetVersionResponse": {
        "properties": {
          "activated_at": {
            "format": "date-time",
            "title": "Activated At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_url",
          "retrieved_at",
          "activated_at",
          "sha256"
        ],
        "title": "ExternalDatasetVersionResponse",
        "type": "object"
      },
      "IdentifierScheme": {
        "enum": [
          "GTIN_8",
          "UPC_A",
          "EAN_13",
          "GTIN_14"
        ],
        "title": "IdentifierScheme",
        "type": "string"
      },
...
      "OpenFoodFactsLookupResponse": {
        "properties": {
          "dataset_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExternalDatasetVersionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "status": {
            "$ref": "#/components/schemas/OpenFoodFactsLookupStatus"
          }
        },
        "required": [
          "status",
          "dataset_version",
          "error_code"
        ],
        "title": "OpenFoodFactsLookupResponse",
        "type": "object"
      },
      "OpenFoodFactsLookupStatus": {
        "enum": [
          "AVAILABLE",
          "NOT_FOUND",
          "UNAVAILABLE"
        ],
        "title": "OpenFoodFactsLookupStatus",
        "type": "string"
      },
      "PackageMatchCandidateResponse": {
        "properties": {
          "allergen_assessment": {
            "$ref": "#/components/schemas/AllergenAssessmentResponse"
          },
          "dataset_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExternalDatasetVersionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "external_record_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Record Id"
          },
          "identity_evidence": {
            "items": {
              "$ref": "#/components/schemas/PackageMatchEvidenceResponse"
            },
            "title": "Identity Evidence",
            "type": "array"
          },
          "label_evidence": {
            "items": {
              "$ref": "#/components/schemas/PackageMatchEvidenceResponse"
            },
            "title": "Label Evidence",
            "type": "array"
          },
          "package_variant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Variant Id"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "reference_images": {
            "items": {
              "$ref": "#/components/schemas/PackageMatchReferenceImageResponse"
            },
            "title": "Reference Images",
            "type": "array"
          },
          "retrieved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieved At"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PackageMatchSourceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_kind": {
            "$ref": "#/components/schemas/PackageMatchSourceKind"
          },
          "source_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Revision"
          }
        },
        "required": [
          "source_kind",
          "allergen_assessment"
        ],
        "title": "PackageMatchCandidateResponse",
        "type": "object"
      },
      "PackageMatchEvidenceResponse": {
        "properties": {
          "dataset_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version Id"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "observed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "source_field": {
            "title": "Source Field",
            "type": "string"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "source_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Revision"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/JsonValue"
          }
        },
        "required": [
          "field",
          "value",
          "source_field",
          "source_name",
          "source_url",
          "language",
          "observed_at",
          "retrieved_at"
        ],
        "title": "PackageMatchEvidenceResponse",
        "type": "object"
      },
      "PackageMatchReferenceImageResponse": {
        "properties": {
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "dataset_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version Id"
          },
          "image_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Revision"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "license_name": {
            "title": "License Name",
            "type": "string"
          },
          "original_url": {
            "title": "Original Url",
            "type": "string"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "source_field": {
            "title": "Source Field",
            "type": "string"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "source_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Revision"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "role",
          "url",
          "original_url",
          "source_field",
          "source_name",
          "source_url",
          "attribution",
          "license_name",
          "language",
          "retrieved_at"
        ],
        "title": "PackageMatchReferenceImageResponse",
        "type": "object"
      },
      "PackageMatchSourceKind": {
        "enum": [
          "REVIEWED_CATALOG",
          "OPEN_FOOD_FACTS"
        ],
        "title": "PackageMatchSourceKind",
        "type": "string"
      },
      "PackageMatchSourceResponse": {
        "properties": {
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "contents_license": {
            "title": "Contents License",
            "type": "string"
          },
          "database_license": {
            "title": "Database License",
            "type": "string"
          },
          "image_license": {
            "title": "Image License",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "record_url": {
            "title": "Record Url",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "terms_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Version"
          }
        },
        "required": [
          "name",
          "source_type",
          "base_url",
          "record_url",
          "attribution",
          "database_license",
          "contents_license",
          "image_license",
          "terms_version"
        ],
        "title": "PackageMatchSourceResponse",
        "type": "object"
      },
      "PackageMatchesResponse": {
        "properties": {
          "candidates": {
            "description": "Candidate Package Matches only; a result does not prove identity with the physical package in a shopper's possession.",
            "items": {
              "$ref": "#/components/schemas/PackageMatchCandidateResponse"
            },
            "title": "Candidates",
            "type": "array"
          },
          "normalized_identifier": {
            "title": "Normalized Identifier",
            "type": "string"
          },
          "open_food_facts": {
            "$ref": "#/components/schemas/OpenFoodFactsLookupResponse"
          },
          "scheme": {
            "$ref": "#/components/schemas/IdentifierScheme"
          }
        },
        "required": [
          "normalized_identifier",
          "scheme",
          "candidates",
          "open_food_facts"
        ],
        "title": "PackageMatchesResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LifeGoods API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/package-matches": {
      "get": {
        "description": "Looks up candidate Package Matches using the active Open Food Facts Dataset Version. Open Food Facts fields are external Evidence, and a returned match does not prove identity with the physical package in a shopper's possession.",
        "operationId": "getPackageMatches",
        "parameters": [
          {
            "description": "GTIN, EAN, or UPC identifier used to find Package Match candidates.",
            "in": "query",
            "name": "identifier",
            "required": true,
            "schema": {
              "description": "GTIN, EAN, or UPC identifier used to find Package Match candidates.",
              "minLength": 1,
              "title": "Identifier",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageMatchesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Find Package Match candidates",
        "tags": [
          "Package Matches"
        ]
      }
    },
...
  }
}